moved logger to utils
