+++ src/modules/__init__.py
@@ -1,4 +1,3 @@
 from .input import Key
 from .config import Config
-from .logger import log
 from .messagebox import showerror
